chore: rename security_svc
