+++ api/api/services/security_service_test.py
@@ -4,7 +4,7 @@ import pytest
 from fastapi import HTTPException
 
 from api.dependencies.security import UserClaims
-from api.services.security_svc import SecurityService
+from api.services.security_service import SecurityService
 from core.domain.errors import InvalidToken
 from core.domain.events import EventRouter
 from core.domain.tenant_data import TenantData
@@ -19,7 +19,7 @@ def mock_org_storage():
 
 
 @pytest.fixture(scope="function")
-def security_svc(mock_org_storage: Mock, mock_event_router: EventRouter):
+def security_service(mock_org_storage: Mock, mock_event_router: EventRouter):
     return SecurityService(mock_org_storage, mock_event_router)
 
 
@@ -59,7 +59,7 @@ class TestUserOrganization:
     async def test_with_api_key(
         self,
         mock_org_storage: Mock,
-        security_svc: SecurityService,
+        security_service: SecurityService,
     ):
         org_settings = TenantData(
             tenant="test_tenant",
@@ -70,7 +70,7 @@ class TestUserOrganization:
         )
         mock_org_storage.find_tenant_for_api_key.return_value = org_settings
 
-        result = await security_svc.find_tenant(None, "wai-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa")
+        result = await security_service.find_tenant(None, "wai-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa")
 
         mock_org_storage.find_tenant_for_api_key.assert_called_once_with(
             "87d3cf28a4ff2a959dff462edff2a40e8bcad5d3a01a6cf2698eac369058075e",
@@ -80,12 +80,12 @@ class TestUserOrganization:
     async def test_with_invalid_api_key(
         self,
         mock_org_storage: Mock,
-        security_svc: SecurityService,
+        security_service: SecurityService,
     ):
         mock_org_storage.find_tenant_for_api_key.side_effect = ObjectNotFoundException()
 
         with pytest.raises(InvalidToken) as e:
-            await security_svc.find_tenant(None, "sk-12345678")
+            await security_service.find_tenant(None, "sk-12345678")
         mock_org_storage.find_tenant_for_api_key.assert_called_once_with(
             "db722c9c5c5fbe334dd0e95a7406d76fcde6396e9d3495edf174ad689fda5dc2",
         )
@@ -98,7 +98,7 @@ class TestUserOrganization:
         self,
         mock_org_storage: Mock,
         new_user: User,
-        security_svc: SecurityService,
+        security_service: SecurityService,
     ):
         org_settings = TenantData(
             tenant="1",
@@ -109,18 +109,18 @@ class TestUserOrganization:
         )
         mock_org_storage.find_tenant_for_org_id.return_value = org_settings
 
-        result = await security_svc.find_tenant(new_user, None)
+        result = await security_service.find_tenant(new_user, None)
 
         mock_org_storage.find_tenant_for_org_id.assert_called_once_with(new_user.org_id)
         assert result == org_settings
 
-    @patch("api.services.security_svc.id_uint32", return_value=1)
+    @patch("api.services.security_service.id_uint32", return_value=1)
     async def test_with_user_org_not_found_creates_new_no_migration(
         self,
         mock_id_uint32: Mock,
         mock_org_storage: Mock,
         new_user: User,
-        security_svc: SecurityService,
+        security_service: SecurityService,
     ):
         """Test the case where the org_id is not found, and the migration fails"""
         mock_org_storage.find_tenant_for_org_id.side_effect = ObjectNotFoundException()
@@ -138,7 +138,7 @@ class TestUserOrganization:
         assert not expected_org.is_anonymous, "sanity, should not be anonymous"
         mock_org_storage.create_organization.return_value = expected_org
 
-        result = await security_svc.find_tenant(new_user, None)
+        result = await security_service.find_tenant(new_user, None)
 
         mock_org_storage.find_tenant_for_org_id.assert_called_once_with(new_user.org_id)
         mock_org_storage.create_organization.assert_called_once_with(expected_org)
@@ -151,13 +151,13 @@ class TestUserOrganization:
         assert result == expected_org
         mock_id_uint32.assert_called_once()
 
-    @patch("api.services.security_svc.id_uint32", return_value=1)
+    @patch("api.services.security_service.id_uint32", return_value=1)
     async def test_with_user_org_not_found_creates_new_anonymous(
         self,
         mock_id_uint32: Mock,
         mock_org_storage: Mock,
         anon_user: User,
-        security_svc: SecurityService,
+        security_service: SecurityService,
     ):
         mock_org_storage.find_anonymous_tenant.side_effect = ObjectNotFoundException()
         expected_org = TenantData(
@@ -170,7 +170,7 @@ class TestUserOrganization:
         mock_org_storage.find_tenant_for_org_id.assert_not_called()
         mock_org_storage.create_organization.return_value = expected_org
 
-        result = await security_svc.find_tenant(anon_user, None)
+        result = await security_service.find_tenant(anon_user, None)
 
         mock_org_storage.find_anonymous_tenant.assert_called_once_with("1234")
         mock_org_storage.create_organization.assert_called_once_with(expected_org)
@@ -181,7 +181,7 @@ class TestUserOrganization:
         self,
         mock_org_storage: Mock,
         anon_user: User,
-        security_svc: SecurityService,
+        security_service: SecurityService,
     ):
         org = TenantData(
             tenant="1234",
@@ -191,7 +191,7 @@ class TestUserOrganization:
         )
         mock_org_storage.find_anonymous_tenant.return_value = org
 
-        result = await security_svc.find_tenant(anon_user, None)
+        result = await security_service.find_tenant(anon_user, None)
 
         mock_org_storage.find_anonymous_tenant.assert_called_once_with("1234")
         mock_org_storage.create_organization.assert_not_called()
@@ -202,7 +202,7 @@ class TestUserOrganization:
         self,
         mock_org_storage: Mock,
         old_user: User,
-        security_svc: SecurityService,
+        security_service: SecurityService,
     ):
         mock_org_storage.find_tenant_for_deprecated_user.return_value = TenantData(
             tenant=old_user.tenant or "",
@@ -214,7 +214,7 @@ class TestUserOrganization:
             slug=old_user.slug or "",
             org_id=old_user.org_id,
         )
-        result = await security_svc.find_tenant(old_user, None)
+        result = await security_service.find_tenant(old_user, None)
         assert result == expected_org
 
         mock_org_storage.find_tenant_for_deprecated_user.assert_called_once_with(
@@ -226,19 +226,19 @@ class TestUserOrganization:
         self,
         mock_org_storage: Mock,
         old_user: User,
-        security_svc: SecurityService,
+        security_service: SecurityService,
     ):
         # Check that we do not autocreate an org if the user token is deprecated
         mock_org_storage.find_tenant_for_deprecated_user.side_effect = ObjectNotFoundException()
 
         with pytest.raises(HTTPException) as e:
-            await security_svc.find_tenant(old_user, None)
+            await security_service.find_tenant(old_user, None)
 
         assert e.value.status_code == 401
         assert e.value.detail == "Organization not found for deprecated token"
 
-    async def test_no_user_no_api_key(self, mock_org_storage: Mock, security_svc: SecurityService):
-        result = await security_svc.find_tenant(None, None)
+    async def test_no_user_no_api_key(self, mock_org_storage: Mock, security_service: SecurityService):
+        result = await security_service.find_tenant(None, None)
 
         mock_org_storage.find_tenant_for_org_id.assert_not_called()
         mock_org_storage.find_tenant_for_api_key.assert_not_called()
@@ -248,7 +248,7 @@ class TestUserOrganization:
         self,
         mock_org_storage: Mock,
         new_user: User,
-        security_svc: SecurityService,
+        security_service: SecurityService,
     ):
         """Test successful migration when finding tenant with org_id"""
         expected_org = TenantData(
@@ -263,7 +263,7 @@ class TestUserOrganization:
         mock_org_storage.find_tenant_for_org_id.side_effect = ObjectNotFoundException()
         mock_org_storage.migrate_tenant_to_organization.return_value = expected_org
 
-        result = await security_svc.find_tenant(new_user, None)
+        result = await security_service.find_tenant(new_user, None)
 
         mock_org_storage.find_tenant_for_org_id.assert_called_once_with(new_user.org_id)
         mock_org_storage.migrate_tenant_to_organization.assert_called_once_with(
@@ -278,7 +278,7 @@ class TestUserOrganization:
         self,
         mock_org_storage: Mock,
         new_user: User,
-        security_svc: SecurityService,
+        security_service: SecurityService,
     ):
         """Test migration failure when finding tenant with org_id"""
         mock_org_storage.find_tenant_for_org_id.side_effect = ObjectNotFoundException()
@@ -294,7 +294,7 @@ class TestUserOrganization:
         )
         mock_org_storage.create_organization.return_value = expected_org
 
-        result = await security_svc.find_tenant(new_user, None)
+        result = await security_service.find_tenant(new_user, None)
 
         mock_org_storage.find_tenant_for_org_id.assert_called_once_with(new_user.org_id)
         mock_org_storage.migrate_tenant_to_organization.assert_called_once_with(
@@ -310,7 +310,7 @@ class TestUserOrganization:
         self,
         mock_org_storage: Mock,
         new_user: User,
-        security_svc: SecurityService,
+        security_service: SecurityService,
     ):
         """Test successful migration when finding tenant with user_id"""
         # Create a user without org_id but with user_id
@@ -334,7 +334,7 @@ class TestUserOrganization:
         mock_org_storage.find_tenant_for_owner_id.side_effect = ObjectNotFoundException()
         mock_org_storage.migrate_tenant_to_user.return_value = expected_org
 
-        result = await security_svc.find_tenant(user_without_org, None)
+        result = await security_service.find_tenant(user_without_org, None)
 
         mock_org_storage.find_tenant_for_owner_id.assert_called_once_with(user_without_org.user_id)
         mock_org_storage.migrate_tenant_to_user.assert_called_once_with(
@@ -344,13 +344,13 @@ class TestUserOrganization:
         )
         assert result == expected_org
 
-    @patch("api.services.security_svc.id_uint32", return_value=1)
+    @patch("api.services.security_service.id_uint32", return_value=1)
     async def test_find_tenant_with_user_id_migration_failure(
         self,
         mock_id_uint32: Mock,
         mock_org_storage: Mock,
         new_user: User,
-        security_svc: SecurityService,
+        security_service: SecurityService,
     ):
         """Test migration failure when finding tenant with user_id"""
         # Create a user without org_id but with user_id
@@ -371,7 +371,7 @@ class TestUserOrganization:
         )
         mock_org_storage.create_organization.return_value = expected_org
 
-        result = await security_svc.find_tenant(new_user, None)
+        result = await security_service.find_tenant(new_user, None)
 
         mock_org_storage.find_tenant_for_owner_id.assert_called_once_with(new_user.user_id)
         mock_org_storage.migrate_tenant_to_user.assert_called_once_with(
